domain -> function 으로 리펙터함
