--- src/main/java/com/googoo/festivaldotcom/domain/festivalApi/service/OpenAiService.java
@@ -1,44 +1,0 @@
-package com.googoo.festivaldotcom.domain.festivalApi.service;
-
-
-import org.json.JSONObject;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.http.*;
-import org.springframework.stereotype.Service;
-import org.springframework.web.client.HttpClientErrorException;
-import org.springframework.web.client.RestTemplate;
-
-@Service
-public class OpenAiService {
-    @Value("${chatgpt.api-key}")
-    private String apiKey;
-
-    private static final String API_URL = "https://api.openai.com/v1/images/generations";
-
-    public String generateImage(String prompt) {
-        RestTemplate restTemplate = new RestTemplate();
-        HttpHeaders headers = new HttpHeaders();
-        headers.setContentType(MediaType.APPLICATION_JSON);
-        headers.setBearerAuth(apiKey);
-
-        JSONObject requestBody = new JSONObject();
-        requestBody.put("prompt", prompt);
-        requestBody.put("n", 1);
-        requestBody.put("size", "1024x1024");
-
-        HttpEntity<String> entity = new HttpEntity<>(requestBody.toString(), headers);
-
-        try {
-            ResponseEntity<String> response = restTemplate.postForEntity(API_URL, entity, String.class);
-            if (response.getStatusCode() == HttpStatus.OK) {
-                JSONObject json = new JSONObject(response.getBody());
-                return json.getJSONArray("data").getJSONObject(0).getString("url");
-            } else {
-                throw new RuntimeException("Failed to generate image: " + response.getStatusCode());
-            }
-        } catch (HttpClientErrorException e) {
-            throw new RuntimeException("API request error: " + e.getStatusCode() + " " + e.getResponseBodyAsString(), e);
-        }
-    }
-
-}
