--- src/main/java/com/googoo/festivaldotcom/domain/member/application/dto/response/UserProfileResponse.java
@@ -1,11 +1,0 @@
-package com.googoo.festivaldotcom.domain.member.application.dto.response;
-
-public record UserProfileResponse(
-	Long id,
-	String nickName,
-	String profileImgUrl,
-	String introduction,
-	Integer festivalCount,
-	String mannerScore
-) {
-}
