+++ src/main/java/com/googoo/festivaldotcom/function/member/application/dto/response/UserProfileResponse.java
@@ -1,0 +1,11 @@
+package com.googoo.festivaldotcom.function.member.application.dto.response;
+
+public record UserProfileResponse(
+	Long id,
+	String nickName,
+	String profileImgUrl,
+	String introduction,
+	Integer festivalCount,
+	String mannerScore
+) {
+}
