+++ src/main/java/com/googoo/festivaldotcom/function/member/infrastructure/repository/UserRepository.java
@@ -1,0 +1,22 @@
+package com.googoo.festivaldotcom.function.member.infrastructure.repository;
+
+import com.googoo.festivaldotcom.function.member.domain.model.User;
+import org.apache.ibatis.annotations.Mapper;
+import org.apache.ibatis.annotations.Param;
+
+import java.util.Optional;
+
+@Mapper
+public interface UserRepository {
+    Optional<User> selectOauthId(String provider, String oauthId);
+
+    Optional<User> selectId(@Param("id") Long id);
+
+    void insertUser(User user);
+
+    int updateUser(@Param("nickName") String nickName, @Param("profileImgUrl") String profileImgUrl, @Param("introduction") String introduction, @Param("id") Long id);
+
+    int deleteUser(@Param("userId") Long userId);
+
+	boolean selectNickName(@Param("nickName") String nickName);
+}
