+++ src/main/resources/application.yml
@@ -86,7 +86,7 @@ sentry:
 # MyBatis
 mybatis:
   mapper-locations: classpath:/mappers/**/*.xml # MyBatis 매퍼 파일의 위치를 설정합니다.
-  type-aliases-package: com.googoo.festivaldotcom.domain # MyBatis에서 사용할 도메인 클래스의 패키지를 설정합니다.
+  type-aliases-package: com.googoo.festivaldotcom.function # MyBatis에서 사용할 도메인 클래스의 패키지를 설정합니다.
   configuration:
     map-underscore-to-camel-case: true # 데이터베이스의 언더스코어 네이밍을 자바의 카멜케이스 네이밍으로 매핑합니다.
