+++ src/main/resources/mappers/fstvData/fstvApiMapper.xml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.googoo.festivaldotcom.domain.festivalApi.repository.FstvlApiRepository">
+<mapper namespace="com.googoo.festivaldotcom.function.festivalApi.repository.FstvlApiRepository">
 
-    <insert id="insertFstvlData" parameterType="com.googoo.festivaldotcom.domain.festivalApi.vo.FstvlData">
+    <insert id="insertFstvlData" parameterType="com.googoo.festivaldotcom.function.festivalApi.vo.FstvlData">
         INSERT INTO FSTVL_DATA (
         FSTVL_NAME, OPAR, FSTVL_START_DATE, FSTVL_END_DATE, FSTVL_CO,
         MNNST_NAME, AUSPC_INSTT_NAME, SUPRT_INSTT_NAME, PHONE_NUMBER,
