+++ src/main/resources/mappers/member/userRepository.xml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.googoo.festivaldotcom.domain.member.infrastructure.repository.UserRepository">
+<mapper namespace="com.googoo.festivaldotcom.function.member.infrastructure.repository.UserRepository">
 
     <!-- Provider와 OAuth ID로 사용자 조회 -->
-    <select id="selectOauthId" resultType="com.googoo.festivaldotcom.domain.member.domain.model.User">
+    <select id="selectOauthId" resultType="com.googoo.festivaldotcom.function.member.domain.model.User">
         SELECT *
         FROM users
         WHERE provider = #{provider}
@@ -13,7 +13,7 @@
     </select>
 
     <!-- 사용자 ID로 사용자 조회 -->
-    <select id="selectId" resultType="com.googoo.festivaldotcom.domain.member.domain.model.User">
+    <select id="selectId" resultType="com.googoo.festivaldotcom.function.member.domain.model.User">
         SELECT *
         FROM users
         WHERE id = #{id}
@@ -21,7 +21,7 @@
     </select>
 
     <!-- 새 사용자 저장 -->
-    <insert id="insertUser" parameterType="com.googoo.festivaldotcom.domain.member.domain.model.User">
+    <insert id="insertUser" parameterType="com.googoo.festivaldotcom.function.member.domain.model.User">
         INSERT INTO users (
             nick_name, introduction, profile_img_url, manner_score,
            festival_count, enabled, provider, oauth_id,
@@ -34,7 +34,7 @@
     </insert>
 
     <!--  사용자 수정  -->
-    <update id="updateUser" parameterType="com.googoo.festivaldotcom.domain.member.application.dto.request.UpdateUserRequest">
+    <update id="updateUser" parameterType="com.googoo.festivaldotcom.function.member.application.dto.request.UpdateUserRequest">
         UPDATE users
         SET nick_name = #{nickName},
             profile_img_url = #{profileImgUrl},
@@ -44,7 +44,7 @@
     </update>
 
     <!--  사용자 삭제  -->
-    <update id="deleteUser" parameterType="com.googoo.festivaldotcom.domain.member.application.dto.request.UpdateUserRequest">
+    <update id="deleteUser" parameterType="com.googoo.festivaldotcom.function.member.application.dto.request.UpdateUserRequest">
         UPDATE users
         SET deleted = 1
         WHERE id = #{id}
